[mcore] Update the resharding logic to per-tensor
