+++ verl/workers/megatron_workers.py
@@ -291,11 +291,11 @@ class ActorRolloutRefWorker(MegatronWorker):
             log_gpu_memory_usage('After building vllm rollout', logger=logger)
 
             # perform weight resharding between actor and rollout
-            sharding_manager = MegatronVLLMShardingManager(
+            sharding_manager = MegatronVLLMShardingManager(module=self.hybrid_engine,
                                                            inference_engine=rollout.inference_engine,
                                                            model_config=self.actor_model_config,
                                                            layer_name_mapping=layer_name_mapping,
-                                                           custom_module=self.actor.actor_module)
+                                                           actor_module=self.actor.actor_module)
             log_gpu_memory_usage('After building sharding manager', logger=logger)
         else:
             raise NotImplementedError('Only vllmRollout is supported with Megatron now')
